Stopped extending E_PMLUnitState
The 'M_' Prefixes annoyed me, and it doesn't seem like anything in the pml library is directly relying on the Unit state ITF. Let me know if you see otherwise. I also removed the subunit info assignment in the fbStateMachine. This link didn't allow me to push a parent into execute without its children. Perhaps with truly strict PML sequence implementation this could make sense, but it seems to be nothing but a restriction at the moment, and a fairly opaque one at thet.