'Improved' event and homing handling
I added some more granularity to a couple of alarms and I attempted to catch reset/error loops, but I don't think I was successful at that yet. I made the base 'Home' just set the fbAxisHome.Execute input, and let the extensions handle the homing behavior by calling the block with their desired parameters. Home to motor torque is not implemented, and home to external feedback is a bit 'Meh'. I don't think the units mean anything, and the particular feedback that I was working with was prone to spikes and inconsistency, so I don't have full confidence in it. Either way, they are some fair placeholders that I hope we can build out better.