Attempted to catch reset loops
I didn't really get to think on this much or prove it out. This may be better handled as a dwell timer that attempts a reset every so often, but that would still be annoying to have new alarms re triggering every couple of seconds. It might be better to have some sort of alarm type within the block that says whether reset should straight up clear the error, clear the execute of the block, or whatever else needs to be done for that particular problem.

Either way, we are going to want 'error' to be hard-driven by present alarms/errors, that way the reset doesn't just magically clear it and allow the events to re-trigger every time and clog ADS.