Big changes to the template
I added initialization placeholders in the commented declaration examples, which I think is super clean and easy to work with compared to past methods of initializing child FBs.
I also added a 'connector' interface array for the Components, just like we have for the Modules, that way, their cyclic calls are handled in the back end and their 'enable' state can be handled better in the event of a parent Abort or reset.. etc

There is also example state machine code in many of the PML state methods, and right now it is presented somewhat like it is the intended operation, so that may need to be packed away a bit more. 

There were certain places that I used the given sequence structure, and it is pretty effective in my opinion, but It was much cleaner to run the machine sequence via PML states and commands.